% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca.R
\name{fpca}
\alias{fpca}
\title{Functional Principal Component Analysis}
\usage{
fpca(ftn, basisname)
}
\arguments{
\item{ftn}{A list containing the coefficients of the functional object and
the basis}

\item{basisname}{A string specifying the type of basis function to use for FPCA}
}
\value{
\code{fpca} returns:
\itemize{
\item \code{pred}: The principal component scores matrix
\item \code{eval}: The eigenvalues
\item \code{mat}: The matrix of eigenfunctions
}
}
\description{
\code{fpca} performs functional principal component analysis (FPCA) on functional
data.
}
\details{
This function performs FPCA on the given functional data and returns the
eigenvalues and eigenfunctions of the covariance matrix.
}
