% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_funct_data.R
\name{gen_funct_data}
\alias{gen_funct_data}
\title{function that generates noiseless functional data using Fourier Basis functions}
\usage{
gen_funct_data(n, p, q, t, eta)
}
\arguments{
\item{n}{Sample size, i.e., the number of observations}

\item{p}{the number of predictors}

\item{q}{the number of Fourier basis functions used to generate the
functional data}

\item{t}{time points}

\item{eta}{A matrix of scores of size pq * pq. These scores can be Gaussian
or elliptical}
}
\value{
\code{gen_funct_data} generates the functional data using Fourier
   basis functions and returns:
   \itemize{
   \item{g: }{the original functional predictors}
   \item{cg: }{the centered functional predictors}
   }
}
\description{
\code{gen_funct_data} generates noiseless functional data using Fourier
basis functions
}
\details{
This function initializes arrays for the original and centered functional
predictors and construct Fourier basis functions over the interval [0, 1].
For each predictor, it computes the functional data by multiplying the
relevant scores with the evaluated basis functions and then centers this data
by subtracting the mean of each time point across all observation.
}
\examples{
# Example 1
n <- 100
p <- 5
q <- 4 (double check!)
t <- seq(0, 1, length.out = 101)
eta <- matrix(rnorm(n * p * q), n, p * q)

result <- gen_funct_data(n, p, q, t, eta)
g <- result$g # Original functional predictors
cg <- result$cg # Centered functional predictors

}
\references{
Wang, G., Liu, S., Han, F., and Di, C.-Z. (2022). Robust functional
   principal component analysis via a functional pairwise spatial sign operator.
   \emph{Biometrics} https://doi.org/10.1111/biom.13695.
}
