% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genfundata.R
\name{genfundata}
\alias{genfundata}
\title{Generate Functional data}
\usage{
genfundata(n, p, nbasis, tt, basisname = "bspline", eta = NULL)
}
\arguments{
\item{n}{Integer. The number of observations, i.e., sample size.}

\item{p}{Integer. The number of functional predictors.}

\item{nbasis}{Integer. The number of basis functions.}

\item{tt}{A numeric vector of time points at which the functional data is
evaluated.}

\item{basisname}{A character string for the type of basis functions to use,
either 'bspline' or 'fourier'.  Default is 'bspline'.}

\item{eta}{An optional array of coefficients of dimension \code{n x p x nbasis}.
If not provided, the coefficients will be randomly generated.}
}
\value{
\code{genfundata} returns a list containing:
\itemize{
\item \code{X}: An array of generated functional data of dimension
\code{n * nt * p}, where \code{nt} denotes the number of time
points.
\item \code{Xc}: An array of centered functional data of dimension
\code{n * nt * p}, where \code{nt} denotes the number of time
points.
\item \code{mfpca.scores}: A matrix of FPCA scores.
.       \item \code{xcoefs}: An array of the coefficients used for generating
the functional data of dimension \code{nbasis * n * p}.
\item \code{xcoefs.mat}: A matrix version of the coefficients with
dimension \code{n * (p * nbasis)}.
\item \code{basis}: The basis object used for functional data generation.
}
}
\description{
\code{genfundata} generates functional data using either B-spline or Fourier
basis functions.
}
\details{
This function constructs functional data using basis expansions, randomly
generates coefficients for the basis functions, applies an orthogonalization
step using a Q-matrix, and performs functional principal component analysis
(FPCA) to obtain a low-dimensional represetation of the data.
}
\examples{
n <- 50
p <- 2
nbasis <- 5
tt <- seq(0, 1, length.out = 100)
data <- genfundata(n, p, nbasis, tt, 'bspline')
str(data)

}
