% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcorr.R
\name{mcorr}
\alias{mcorr}
\title{Multiple correlation}
\usage{
mcorr(u, v)
}
\arguments{
\item{u, v}{Numeric vectors or matrices.}
}
\value{
\code{mcorr} returns a numeric value representing the multiple
correlation between \code{u} and \code{v}.
}
\description{
\code{mcorr} computes the multiple correlation between two numeric matrices
\code{u} and \code{v} of the same dimension.
}
\details{
This function computes the multiple correlation between two matrices
\code{u} and \code{v} of the same dimension.  If \code{u} and \code{v}
are column vectors, then \code{mcorr} is equivalent to the \code{cor} function
from the \code{stats} package.

Let \eqn{C_{uu}}, \eqn{C_{uv}}, \eqn{C_{vu}}, and \eqn{C_{vv}} represent
the sample covariance matrices. The multiple correlation between \code{u}
and \code{v} is given by:
\deqn{
mcorr(u, v) = tr(C_{vv}^{-1/2} C_{vu} C_{uu}^{-1} C_{uv} C_{vv}^{-1/2}).
}
This number ranges from 0 to d, where d denotes the number of columns in
\code{u} and \code{v}. A number closer to \code{d} indicates stronger correlation.
This measure has been used in functional dataaa dimension reduction studies,
such as Li and Song (2022) and Solea et al. (2024).
}
\examples{
# Example 1: Strong correlation
u <- matrix(rnorm(100), ncol = 2)
v <- u + 0.1 * matrix(rnorm(100), ncol = 2)
mcorr(u, v)

# Example 2: Weak correlation
u <- matrix(rnorm(150), ncol = 3)
v <- matrix(rnorm(150), ncol = 3)
mcorr(u, v)

# Example 3: Comparison with Pearson correlation
u <- rnorm(100)
v <- u + 0.1 * rnorm(100)
mcorr(u, v)
stats::cor(u, v)

}
\references{
Li, B, and Song, J. (2022). Dimension reduction for functional
data based on weak conditional moments. \emph{The Annals of Statistics}
50(1), 107--128.

Solea, E., Christou, E., and Song, J. (2024). Robust Inverse Regression for
Multivariate Elliptical Functional Data. \emph{Statistica Sinica}
https://doi.org/10.5705/ss.202023.0341
}
\seealso{
\code{\link[stats]{cor}}, \code{\link[stats]{cov}}, \code{\link[stats]{var}}
}
