% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcqs.R
\name{fcqs}
\alias{fcqs}
\title{Functional Central Quantile Subspace}
\usage{
fcqs(x, y, tt, tau = 0.5, dtau = NULL, nbasis = 4, norder = 4)
}
\arguments{
\item{x}{A 3-dimensional array (\code{n x nt x p}), where n is the number
of observations, nt is the number of time points, and p is the number
of predictor variables.}

\item{y}{A numeric vector of length \code{n} representing the scalar
response variable.}

\item{tt}{A numeric vector of length \code{nt} of time points at which
the functional data is evaluated.}

\item{tau}{A quantile level, a number strictly between 0 and 1.}

\item{dtau}{The number of directions to extract, i.e., the dimension of
the functional central quantile subspace.  It should be an integer
between 1 and \code{p}.  If not provided, the function will return \code{p}
directions}

\item{nbasis}{An integer specifying the number of basis functions for
functional smoothing.  Currently, the only option is to use B-spline
basis and the default value is 4.}

\item{norder}{An integer specifying the order of B-splines, which is one
higher than their degree.  The default of 4 gives cubic splines.}
}
\value{
\code{fcqs} computes the directions of the functional central quantile
subspace (FCQS) and returns:
\item{ffun}{A matrix of size \code{nbasis x dtau}, representing
the functional parameters that span the FCQS}
\item{betax}{A matrix of size \code{n x dtau}, representing the
sufficient predictor, calculated as the inner product between
\code{betacoef} and \code{x}.}
}
\description{
\code{fcqs} estimates the directions of the functional central quantile
subspace.
}
\details{
This function computes the directions that span the \eqn{\tau}th functional
central quantile subspace. These directions represent functions that can
be linearly applied via the inner product to replace the infinite-dimensional
functional predictors with a few finite predictors without losing important
information on the conditional quantiles while maintaining a flexible
nonparametric model. This methodology is introduced in Christou et al. (2025+).
}
\examples{
# Set the parameters
n <- 100
p <- 5
nbasis <- 4
nt <- 100
tau <- 0.1
dtau <- 1
tt <- seq(0, 1, length.out = nt)
# Set the covariance matrix
SigmaCov <- matrix(0, p * nbasis, p * nbasis)
for (j in 1:p) {
 index.j <-(((j - 1) * nbasis + 1):(j * nbasis))
 diag(SigmaCov[index.j, index.j]) <- c(2, 1, 1/2, 1/4)
}
eta.mat <- mvtnorm::rmvnorm(n, mean = rep(0, p * nbasis), sigma = SigmaCov)
eta <- array(eta.mat, dim = c(nbasis, n, p))
# Generate functional data
data <- fundata(n, p, nbasis, tt, 'fourier', eta)
Xc <- data$Xc
# Generate y
error <- rnorm(n)
y <- 3 * data$mfpca.scores[, 1] + error
# Run FCQS function
output <- fcqs(Xc, y, tt, tau, dtau)
# Compare the true and estimated predictors
mcorr(output$betax, data$mfpca.scores[, 1])
}
\references{
Christou, E., Solea, E., Wang, S., and Song, J. (2025+) Sufficient
Dimension Reduction for the Conditional Quantiles of Functional Data.
\emph{under revision}.
}
