% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcqs.R
\name{fcqs}
\alias{fcqs}
\title{Functional Central Quantile Subspace}
\usage{
fcqs(x, y, time, nbasis, tau = 0.5, d_tau)
}
\arguments{
\item{x}{A 3-dimensional array (\code{n x nt x p}), where n is the number
of observations, nt is the number of time points, and p is the number
of predictor variables.}

\item{y}{A numeric vector of length \code{n} representing the response
variable.}

\item{time}{A numeric vector of length \code{nt} of time points at which
the functional data is evaluated.}

\item{nbasis}{The number of basis functions for smoothing functional data.}

\item{tau}{A quantile level, a number strictly between 0 and 1.}

\item{d_tau}{The dimension of the functional central quantile subspace.
It should be an integer between 1 and \code{p}.}
}
\value{
\code{fcqs} computes the directions of the functional central quantile
subspace (FCQS) and returns:
\item{betacoef}{The functional parameters that span the FCQS}
\item{yhat}{The estimated responses, resulting as the inner product
between \code{betacoef} and \code{x}.}
}
\description{
\code{fcqs} estimates the directions of the functional central quantile
subspace, extending the central quantile subspace to functional data.
}
\details{
This function computes the directions that span the \eqn{\tau}th functional
central quantile subspace. These directions represent functions that can
be linearly applied via the inner product to given predictors to reduce the
dimension of infinitely-dimensional functional data without losing any
information required for accurate quantile regression of the functional
data.
}
\examples{
# Parameters for generating functional data
n <- 100
p <- 5
q <- 4
time <- seq(0, 1, length.out = 101)
eta <- matrix(stats::rnorm(n * p * q), nrow = n, ncol = p * q)

# Generate functional data
result <- fundata(n, p, q, time, eta)

# Centered functional predictors
xc <- result$xc

# Further parameters for FCQS
H <- 10
tau <- 0.1
K <- 1
d_tau <- 1
nbasis <- q
d_DR <- 1
P <- eigen(cov(eta))$vectors
Q <- diag(eigen(cov(eta))$values)
# This is the inner products <b1, X>, <b2, X>, ...
mfpca.scores <- eta \%*\% P
# Define true inner products for functional central subspace
uu_fcs <- mfpca.scores[, 1:K]
# Generate error for data
error <- rnorm(n)
# Generate response from functional data and error
y <- 3 * mfpca.scores[, 1] + error
# Run FCQS function
fcqs(xc, y, time_points, q, nbasis, tau, d_tau, H, d_DR)
}
