% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcqs.R
\name{fcqs}
\alias{fcqs}
\title{Functional Central Quantile Subspace}
\usage{
fcqs(x, y, time, nbasis, tau = 0.5, d_tau)
}
\arguments{
\item{x}{A 3-dimensional array (\code{n x nt x p}), where n is the number
of observations, nt is the number of time points, and p is the number
of predictor variables.}

\item{y}{A numeric vector of length \code{n} representing the response
variable.}

\item{time}{A numeric vector of length \code{nt} of time points at which
the functional data is evaluated.}

\item{nbasis}{The number of basis functions for smoothing functional data.}

\item{tau}{A quantile level, a number strictly between 0 and 1.}

\item{d_tau}{The dimension of the functional central quantile subspace.
It should be an integer between 1 and \code{p}.}
}
\value{
\code{fcqs} computes the directions of the functional central quantile
subspace (FCQS) and returns:
\item{betacoef}{The functional parameters that span the FCQS}
\item{betax}{The resulting sufficient predictor, calculated as the
inner product between \code{betacoef} and \code{x}.}
}
\description{
\code{fcqs} estimates the directions of the functional central quantile
subspace.
}
\details{
This function computes the directions that span the \eqn{\tau}th functional
central quantile subspace. These directions represent functions that can
be linearly applied via the inner product to given predictors to reduce the
dimension of infinitely-dimensional functional data without losing any
information required for accurate quantile regression of the functional
data.
}
\examples{
# Set the parameters
n <- 100
p <- 5
nbasis <- 4
tau <- 0.1
d_tau <- 1
time <- seq(0, 1, length.out = 101)
eta <- matrix(stats::rnorm(n * p * nbasis), nrow = n, ncol = p * nbasis)
# Generate functional data
result <- fundata(n, p, nbasis, time, eta)
xc <- result$xc
# Generate y
P <- eigen(cov(eta))$vectors
mfpca.scores <- eta \%*\% P
error <- rnorm(n)
y <- 3 * mfpca.scores[, 1] + error
# Run FCQS function
output <- fcqs(xc, y, time, nbasis, tau, d_tau)
# Compare the true and estimated predictors
mcorr(output$betax, mfpca.scores[, 1])
}
