% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcorr.R
\name{mcorr}
\alias{mcorr}
\title{multiple correlation}
\usage{
mcorr(u, v)
}
\arguments{
\item{u}{A numeric vector or matrix. The first variable for the correlation
measure}

\item{v}{A numeric vector or matrix. The second variable for the correlation
measure}
}
\value{
\code{mcorr} computes the multiple correlation between \code{u}
and \code{v}
}
\description{
\code{mcorr} computes the multiple correlation between two random vectors
\code{u} and \code{v}
}
\details{
This function computes the multiple correlation between two random vectors
\code{u} and \code{v} of the same dimension. Let C_{uu}, C_{uv}, and C_{vu}
represent the sample covariance matrices, the multiple correlation between
\code{u} and \code{v} is
\deqn{
mcorr(u, v) = tr(C_{vv}^{-1/2} C_{vu} C_{uu}^{-1} C_{uv} C_{vv}^{-1/2})
}
}
\examples{
set.seed(1234)
u <- matrix(rnorm(100), ncol=2)
v <- matrix(rnorm(100), ncol=2)
mcorr(u, v)

}
\references{
Li, B, and Song, J. (2022). Dimension reduction for functional
data based on weak conditional moments. \emph{The Annals of Statistics}
50(1), 107--128.
}
