% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sonf.R
\name{sonf}
\alias{sonf}
\title{Scalar-on-function regression}
\usage{
sonf(y, xfd, dev2_penalty = FALSE, lambda = NULL)
}
\arguments{
\item{y}{A number vector of length n, representing the scalar response variable}

\item{xfd}{A functional data object}

\item{dev2_penalty}{A logical flag indicating whether to apply a second
derivative penalty.}

\item{lambda}{A penalty parameter (used if \code{dev2_penalty} is 'TRUE')}
}
\value{
\code{sonf} computes a scalar-on-function regression and returns:
\itemize{
\item \code{beta_coef}: The regression coefficients.

\item \code{beta_fd}: The functional data objects for the beta coefficients.

\item \code{yhat}: The predicted values of \code{y}.

\item \code{X}: The designed matrix used in the regression.
}
}
\description{
\code{sonf} estimates the relationship between a scalar response and
functional predictors using basis functions, with options for applying
a second derivative penalty or not. Test
}
\details{
This function performs scalar-on-function regression, predicting a scalar
response \code{y} using one or more functional predictors \code{xfd}. It
centralizes the data, constructs a design matrix based on the different type
of basis functions, and calculates the regression coefficients with or without
a second derivative penalty. The function returns the regression coefficients,
functional data objects for these coefficients, estimated response value,
and the designed matrix X.
}
